+++ src/amd.jquery.js
@@ -19,7 +19,8 @@
         , currentLng
         , replacementCounter = 0
         , languages = []
-        , initialized = false;
+        , initialized = false
+        , sync = {};
 
     //= i18next.sync.js
     //= i18next.defaults.js